refactor: modularize src into domain folders with distributed tests
Split the monolithic provider.ts, extension.ts, and utils.ts into focused
domain modules under src/extension/, src/provider/, and src/shared/.
Distribute tests to mirror the source structure under src/test/.

- extension/: commands, diagnostics, serverManagement, serverRegistry, status
- provider/: client, config, discovery, modelDefaults, registration, request, streaming
- shared/: json, messages, tools, validation
- provider.ts reduced to pure orchestration (~240 lines)
- utils.ts removed; imports go directly to shared modules
- Add zod for boundary validation schemas in discovery
- 140 tests passing
