+++ package.json
@@ -142,5 +142,8 @@
 	"packageManager": "bun@1.3.9",
 	"lint-staged": {
 		"*.{js,ts,json,md,yml,yaml}": "prettier --write"
+	},
+	"dependencies": {
+		"zod": "^4.4.3"
 	}
 }
