+++ src/extension/serverRegistry.ts
