+++ src/shared/json.ts
@@ -1,0 +1,19 @@
+/**
+ * Try to parse a JSON object from a string.
+ * @param text The input string.
+ * @returns Parsed object or ok:false.
+ */
+export function tryParseJSONObject(text: string): { ok: true; value: Record<string, unknown> } | { ok: false } {
+	try {
+		if (!text || !/[{]/.test(text)) {
+			return { ok: false };
+		}
+		const value = JSON.parse(text);
+		if (value && typeof value === "object" && !Array.isArray(value)) {
+			return { ok: true, value };
+		}
+		return { ok: false };
+	} catch {
+		return { ok: false };
+	}
+}
