+++ src/shared/validation.ts
@@ -1,0 +1,52 @@
+import * as vscode from "vscode";
+import { isToolResultPart } from "./messages";
+
+/**
+ * Validate the request message sequence for correct tool call/result pairing.
+ */
+export function validateRequest(messages: readonly vscode.LanguageModelChatRequestMessage[]): void {
+	const lastMessage = messages[messages.length - 1];
+	if (!lastMessage) {
+		console.error("[LiteLLM Model Provider] No messages in request");
+		throw new Error("Invalid request: no messages.");
+	}
+
+	messages.forEach((message, i) => {
+		if (message.role === vscode.LanguageModelChatMessageRole.Assistant) {
+			const toolCallIds = new Set(
+				message.content
+					.filter((part) => part instanceof vscode.LanguageModelToolCallPart)
+					.map((part) => (part as unknown as vscode.LanguageModelToolCallPart).callId)
+			);
+			if (toolCallIds.size === 0) {
+				return;
+			}
+
+			let nextMessageIdx = i + 1;
+			const errMsg =
+				"Invalid request: Tool call part must be followed by a User message with a LanguageModelToolResultPart with a matching callId.";
+			while (toolCallIds.size > 0) {
+				const nextMessage = messages[nextMessageIdx++];
+				if (!nextMessage || nextMessage.role !== vscode.LanguageModelChatMessageRole.User) {
+					console.error(
+						"[LiteLLM Model Provider] Validation failed: missing tool result for call IDs:",
+						Array.from(toolCallIds)
+					);
+					throw new Error(errMsg);
+				}
+
+				nextMessage.content.forEach((part) => {
+					if (!isToolResultPart(part)) {
+						const ctorName =
+							(Object.getPrototypeOf(part as object) as { constructor?: { name?: string } } | undefined)?.constructor
+								?.name ?? typeof part;
+						console.error("[LiteLLM Model Provider] Validation failed: expected tool result part, got:", ctorName);
+						throw new Error(errMsg);
+					}
+					const callId = (part as { callId: string }).callId;
+					toolCallIds.delete(callId);
+				});
+			}
+		}
+	});
+}
